Extract PartitionSet and related types to own module.
